populate initial keyring
traditional installs use pacstrap without the -G flag, which causees the
host's keyring (ie the live CD) to be imported onto the guest. this
provides an initial pacman keyring for archlinux-keyring to use, which
avoids the need to populate the master keys manually.

when you pass -G to pacstrap, the host keyring is not imported to the
guest, and the guest has no keys at all. installing archlinux-keyring is
supposed to populate the master keys, but if pacman is not installed at
that time (as when the system is being bootstrapped), then the
pacman-key binary does not exist, so keyring initialization is
impossible.

to fix this, we have to manually populate the keyring after installing
pacman. (note that --init creates the user's key on the pacman keyring,
which is used to sign the master keys imported by --populate. those two
commands are not equivalent.)