add note about /dev/fd at build time
somehow, this symlink already exists during acbuild run. i can't find
references to it in the code, so i assume it comes in from nspawn
somehow. maybe because stdout/stdin are a tty at build time? anyway,
since i can't put the symlink into the aci, it actually has to be made
in the run script