inherit host resolv.conf
so i encountered an interesting bug while rebuilding my base image
today: it didn't have a resolv.conf in it. which in and of itself is
perhaps unsurprising, but an image i built a few days ago (2017-10-25)
had the standard resolvconf-generated resolv.conf baked into its aci
(which is super weird, because openresolv wasn't installed in that image
at any point).

anyway, the solution to this is to bind mount the host's resolv.conf in,
since we're already sharing the rest of the network